+++ Changelog.md
@@ -7,6 +7,9 @@ Breaking Changes for 3.0.0:
 * Remove the deprecated `be_close` matcher, preferring `be_within` instead.
   (Sam Phippen)
 * Rename `be_true` and `be_false` to `be_truthy` and `be_falsey`. (Sam Phippen)
+* Make `expect { }.to_not raise_error(SomeSpecificClass, message)`,
+       `expect { }.to_not raise_error(SomeSpecificClass)` and
+       `expect { }.to_not raise_error(message)` invalid. (Sam Phippen)
 
 Enhancements: